+++ main.h
@@ -27,7 +27,7 @@ int _putchar(char c);
 int _puts(char *str);
 int _strlen(char *s);
 int _printf(const char *format, ...);
-int _putnbr(int n, t_flag flags);
+int _putnbr(long int n, t_flag flags);
 char *_uitoa_base(unsigned long int n, int base);
 int _isinstr(char *str, char toFind);
 void _to_upper(char *s);
